+++ android/expoview/src/main/java/host/exp/exponent/experience/HomeActivity.kt
@@ -26,7 +26,6 @@ import expo.modules.easclient.EASClientModule
 import expo.modules.facedetector.FaceDetectorPackage
 import expo.modules.filesystem.FileSystemModule
 import expo.modules.filesystem.FileSystemPackage
-import expo.modules.font.FontLoaderPackage
 import expo.modules.haptics.HapticsModule
 import expo.modules.keepawake.KeepAwakeModule
 import expo.modules.keepawake.KeepAwakePackage
@@ -157,7 +156,6 @@ open class HomeActivity : BaseExperienceActivity() {
         ConstantsPackage(),
         PermissionsPackage(),
         FileSystemPackage(),
-        FontLoaderPackage(),
         BarCodeScannerPackage(),
         KeepAwakePackage(),
         FaceDetectorPackage(),